fix: make envs own functions as well
Previously, if an env were to be used to compile a single program, then
said program were to be destroyed, custom functions registered to that
env would be destroyed, which is incorrect for obvious reasons.

This diff makes custom functions start with a reference count of 1,
representing the environment they were registered to. The environment
destruction logic is also adjusted accordingly.